Remove fake WorkStore around unstable_cache in Pages (#70809)
This was a hack that effectively put us inside App Router mode when
`unstable_cache` was used inside Pages.

I removed the need for this to shadow the global `force-no-store` option
in https://github.com/vercel/next.js/pull/70740 and it should no longer
be needed for shadowing mutations on the WorkStore object. E.g. tags
collected by the inner cache.

In #70808 we also don't need this to put us inside App Router mode when
`unstable_cache` was used inside Middleware.