Add pathname to PrerenderStore (#70889)
This lets us use this for implicit tags when a full Request is not
available. Required for #70819.

This should probably be more normalized but I only pass the path name
instead of the full url to avoid leaking more dynamic information to the
prerender.

This does not yet solve the case where we have nested caches (which we
don't in unstable_cache but in use cache). In that case it's more
complex because for an outer hit we'd need to add it to the inner cache
entries.