Update 04-linking-and-navigating.mdx, Partial Rendering, explain about unmounting and losing state (#70397)
### What?

Update 04-linking-and-navigating.mdx, 4. Partial Rendering section, add
information that rerenders of sibling routes mean full rerender of the
page, unmounting and losing the state of the page

### Why?
I've spent a lot of time trying to understand why in the app router when
switching between sibling routes (and Dynamic Segments) I lose scroll
position. Finally, find a discussion that this is the intentional
behaviour of the Next.js team. Would be good if the Next.js users could
find information in the documentation instead of searching for issues
and discussions in the documentation.

https://github.com/vercel/next.js/discussions/50711

### How?
Updated docs

---------

Co-authored-by: JJ Kasper <jj@jjsweb.site>