Run after() callbacks in the same context as the outer function (#70810)
Stacked on #70806.

We want `after` callbacks to run in the same ALS context that `after`
was called in. This can be achieved via
https://nodejs.org/api/async_context.html#static-method-asynclocalstoragebindfn

We want this be cause these should be as close as possible in every way
(except timing):
```
after(() => x())
after(x())
await x()
```

We'll need a different approach for disabling setting cookies (and
revalidate). This will be done in a follow up.

---------

Co-authored-by: Janka Uryga <lolzatu2@gmail.com>