Skip `Math.random` test with route handler and `'use cache'` (#71025)
Using `'use cache'` in route handlers is not yet fully supported, and
fails in isolation, e.g. in a deployment test, because it implicitly
relies on the fact that a page must have been compiled before the route.
Otherwise rendering the route errors with "Missing manifest for Server
Actions.".

This deploy test failure slipped through because #70837 was created from
a fork.